pyproject.toml
[tool.poetry]
name = "pytorch_prednet"
version = "0.1.0"
description = ""
authors = ["javiferfer <javier.f3rnand3z@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.9"
torch = "^1.12.1"
torchvision = "^0.13.1"
Pillow = "^8.0.1"
torchviz = "^0.0.1"
tqdm = "^4.56.0"
tensorboard = "^2.4.1"
librosa = "^0.8.0"
matplotlib = "^3.4.1"
SoundFile = "^0.10.3"
opencv-python = "^4.5.1"
h5py = "^3.7.0"
numpy = "^1.21.3"
scipy = "^1.7.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
